# intellij-projects/functional-test-rest-service/api-docs.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Profile Component",
    "description": "A dec",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8585",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/profile/": {
      "post": {
        "tags": [
          "profile-controller"
        ],
        "operationId": "createProfile",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserProfile"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfile"
                }
              }
            }
          }
        }
      }
    },
    "/profile/{id}": {
      "get": {
        "tags": [
          "profile-controller"
        ],
        "operationId": "getProfile",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfile"
                }
              }
            }
          }
        }
      }
    },
    "/profile/users": {
      "get": {
        "tags": [
          "profile-controller"
        ],
        "operationId": "getAllProfiles",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserProfile"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "UserProfile": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          }
        }
      }
    }
  }
}